move components to subdir
